Module glslt::transform

GLSLT transform utilities definitions

Modules

template
Definitions of template function abstractrepresentations

Structs

FnRef
Reference to a function held in a transformation unit

MinUnit
Minimizing GLSLT template instantiation context

Unit
GLSLT template instantiation context

Enums

FnHandle
Handle to a function definition somewhere in the glslt units

Traits

TransformUnit
Represent a GLSLT transform unit

Functions

transform
Transform a GLSLT AST to an instantiated GLSL AST

transform_min
Transform a GLSLT AST to an instantiated GLSL AST. Only include symbols transitively referenced by any of the entry points listed in wanted.